Make Entity handle the display card body text
Originally the OwnerCard will manually retrieve each property (address, email, etc) manually. As a lot of these properties are specific to the type of Entity, this is not ideal for abstraction purposes. Thus it is necessary to let Entity class be able to return the properties and reduce reliance on the specific classes.