Integrate ID into application
The ID is unique across all entities since all of them are stored in the
same container and thus naturally they will not clash. Current
implementation is by using a list of pairs. This allows for the UI
portion of the code to be adapted to use the new ID instead of the
visible list index.

Future work will include recording the list of dog owned by one single
owner. This will also necessitate a check right after import from any
other list. Tests coverage also dropped by a lot, there is huge amount
of test code that can be updated or newly written.